src/requirements/base.txt
# Core
Django>=1.8,<1.9
django-crispy-forms==1.4
django-markdown-deux==1.0.5
psycopg2>=2.6,<2.7
social-auth-core==1.2
social-auth-app-django==1.1

# Packages
ipython
django-extensions
pillow
Werkzeug
